split up gql srings into multiple files
