+++ frontend/src/pages/Programs/ProgramPage/index.tsx
@@ -16,7 +16,7 @@ import ProgramDiscussions from "./Discussions";
 import graphQLClient from "@/apiInterface/client";
 
 import "./index.css";
-import { JOIN_PROGRAM_MUTATION } from "@/apiInterface/gqlOperations";
+import { JOIN_PROGRAM_MUTATION } from "@/apiInterface/gqlStrings/programStrings";
 import { subscribeToUserModelSubject } from "@/observers/userobserver";
 import { User } from "@/models/usermodel";
 import { ProgramContext } from "./ProgramContext";
